+++ src/partials/header.html
@@ -76,7 +76,7 @@
   <div class="backdrop js-menu-container is-hidden">
     <div class="mob-menu">
       <div class="mob-menu__header">
-        <a href="#" class="logo" id="myElement">News</a>
+        <a href="#" class="logo mob-logo" id="myElement">News</a>
         <svg class="mob-menu__x js-close-menu" width="24" height="24">
           <use href="./images/sprite-full.svg#icon-close"></use>
         </svg>
@@ -87,7 +87,7 @@
             ><svg class="mob-menu__icon" width="18" height="18">
               <use href="./images/sprite-full.svg#icon-home"></use>
             </svg>
-            <p>Home</p>
+            <p class="mob-link-text">Home</p>
             <svg class="mob-menu__icon-arrow" width="18" height="18">
               <use href="./images/sprite-full.svg#icon-arrow-right"></use></svg
           ></a>
@@ -100,7 +100,7 @@
             ><svg class="mob-menu__icon" width="18" height="18">
               <use href="./images/sprite-full.svg#icon-favorite"></use>
             </svg>
-            <p>Favorite</p>
+            <p class="mob-link-text">Favorite</p>
             <svg class="mob-menu__icon-arrow" width="18" height="18">
               <use href="./images/sprite-full.svg#icon-arrow-right"></use></svg
           ></a>
@@ -110,7 +110,7 @@
             ><svg class="mob-menu__icon" width="18" height="18">
               <use href="./images/sprite-full.svg#icon-read"></use>
             </svg>
-            <p>Read</p>
+            <p class="mob-link-text">Read</p>
             <svg class="mob-menu__icon-arrow" width="18" height="18">
               <use href="./images/sprite-full.svg#icon-arrow-right"></use>
             </svg>
